with external feats
