+++ sm_cnn/args.py
@@ -22,7 +22,8 @@ def get_args():
     parser.add_argument('--words_dim', type=int, default=50)
     parser.add_argument('--dropout', type=float, default=0.5)
     parser.add_argument('--epoch_decay', type=int, default=15)
-    parser.add_argument('--vector_cache', type=str, default="data/word2vec.trecqa.pt")
+    parser.add_argument('--vector_cache', type=str, default='data/word2vec.trecqa.pt')
+    parser.add_argument('--external_cache', type=str, default='data/overlap_feats.pt')
     parser.add_argument('--trained_model', type=str, default="")
     parser.add_argument('--weight_decay',type=float, default=0)
     parser.add_argument('--use_ext', action='store_true')
